+++ src/klarray.h
@@ -11,7 +11,7 @@ extern "C" {
 /**
  * Creates a new array.
  */
-extern KLArray *kl_array_new(UInt64 elemSize, UInt64 capacity);
+extern KLArray *kl_array_new(size_t elemSize, size_t capacity);
 
 /**
  * Destroys an array.
@@ -21,28 +21,28 @@ extern void kl_array_destroy(KLArray *array);
 /**
  * Gets the number of elements in the array.
  */
-extern UInt64 kl_array_count(KLArray *array);
+extern size_t kl_array_count(KLArray *array);
 
 /**
  * Gets the capacity of the array (in number of elements).
  */
-extern UInt64 kl_array_capacity(KLArray *array);
+extern size_t kl_array_capacity(KLArray *array);
 
 /**
  * Ensures that there are enough space for the extra number of elements in the
  * array.
  */
-extern bool kl_array_ensure_capacity(KLArray *array, UInt64 extraCapacity);
+extern bool kl_array_ensure_capacity(KLArray *array, size_t extraCapacity);
 
 /**
  * Returns the size of each element in the array
  */
-extern UInt64 kl_array_element_size(KLArray *array);
+extern size_t kl_array_element_size(KLArray *array);
 
 /**
  * Returns a pointer to the element at a particular index in the array.
  */
-extern void *kl_array_element_at(KLArray *array, Int64 index);
+extern void *kl_array_element_at(KLArray *array, off_t index);
 
 /**
  * Returns a random element.
@@ -52,13 +52,13 @@ extern void *kl_array_random(KLArray *array);
 /**
  * Removes an element at a given index.
  */
-extern void kl_array_remove_at(KLArray *array, Int64 index);
+extern void kl_array_remove_at(KLArray *array, off_t index);
 
 /**
  * Inserts a value at a given index and returns the pointer to it 
  * so the caller can fill it as they see fit.
  */
-extern void *kl_array_insert_at(KLArray *array, Int64 index);
+extern void *kl_array_insert_at(KLArray *array, off_t index);
 
 /**
  * Returns a pointer to the first element in the array.
