+++ src/klbuffer.h
@@ -11,7 +11,7 @@ extern "C" {
 /**
  * Creates a new buffer with a given capacity.
  */
-extern KLBuffer *kl_buffer_new(UInt64 capacity);
+extern KLBuffer *kl_buffer_new(size_t capacity);
 
 /**
  * Destroys an buffer.
@@ -21,18 +21,18 @@ extern void kl_buffer_destroy(KLBuffer *buffer);
 /**
  * Gets the number of bytes in the buffer.
  */
-extern UInt64 kl_buffer_size(KLBuffer *buffer);
+extern size_t kl_buffer_size(KLBuffer *buffer);
 
 /**
  * Gets the available capacity of the buffer.
  */
-extern UInt64 kl_buffer_capacity(KLBuffer *buffer);
+extern size_t kl_buffer_capacity(KLBuffer *buffer);
 
 /**
  * Ensures that there are enough space for the extra number of elements in the
  * buffer.
  */
-extern bool kl_buffer_ensure_capacity(KLBuffer *buffer, UInt64 extraBytes);
+extern bool kl_buffer_ensure_capacity(KLBuffer *buffer, size_t extraBytes);
 
 /**
  * Returns the underlying bytes of the buffer.
@@ -42,13 +42,13 @@ extern char *kl_buffer_bytes(KLBuffer *buffer);
 /**
  * Appends data to the end of the buffer.
  */
-extern void kl_buffer_append(KLBuffer *buffer, const char *data, UInt64 datasize);
+extern void kl_buffer_append(KLBuffer *buffer, const char *data, size_t datasize);
 
 /**
  * Copy a portion of the buffer to the output buffer and returns the number of
  * bytes copied.
  */
-extern UInt64 kl_buffer_copy(KLBuffer *buffer, Int64 srcOffset, char *output, UInt64 numBytes);
+extern size_t kl_buffer_copy(KLBuffer *buffer, Int64 srcOffset, char *output, size_t numBytes);
 
 /**
  * Clears the buffer and resets to the position to the start.
